Added binance provider, refactored code
