+++ .env
@@ -3,6 +3,7 @@ PORT=2777
 COINGECKO_URL=https://api.coingecko.com/api/v3/simple/price?ids=%s&vs_currencies=%s
 COINAPI_URL=https://rest.coinapi.io/v1/exchangerate/%v/%v
 COINAPI_KEY=C72B3E65-2D6F-4AFC-AA94-AD84537D1825
+BINANCEAPI_URL=https://api.binance.com/api/v3/ticker/price?symbol=%v%v
 CURRENCY_FROM=BTC
 CURRENCY_TO=UAH
