+++ internal/configuration/logger/logger.go
@@ -1,7 +1,7 @@
 package logger
 
 import (
-	"btc-test-task/internal/helpers/config"
+	"btc-test-task/internal/configuration/config"
 	"bytes"
 	"fmt"
 	"os"
