+++ internal/currencyrate/coin_gecko_executor.go
@@ -1,9 +1,8 @@
-package rateProviders
+package currencyrate
 
 import (
-	"btc-test-task/internal/helpers/config"
-	"btc-test-task/internal/helpers/customErrors"
-	"btc-test-task/internal/helpers/models"
+	"btc-test-task/internal/configuration/config"
+	models "btc-test-task/internal/models"
 	"bytes"
 	"encoding/json"
 	"fmt"
@@ -29,10 +28,6 @@ func NewCoinGeckoExecutor(conf *config.Config) *CoinGeckoExecutor {
 	}
 }
 
-func (api *CoinGeckoExecutor) generateEndpoint(currency *models.Currency) string {
-	return fmt.Sprintf(api.url, currency.CurrencyFrom, currency.CurrencyTo)
-}
-
 func (api *CoinGeckoExecutor) GenerateHttpRequest(currency *models.Currency) (*http.Request, error) {
 	endpoint := api.generateEndpoint(currency)
 	req, err := http.NewRequest(
@@ -40,7 +35,7 @@ func (api *CoinGeckoExecutor) GenerateHttpRequest(currency *models.Currency) (*h
 		endpoint,
 		nil)
 	if err != nil {
-		return nil, errors.Wrap(err, "GenerateCurrencyRequest")
+		return nil, errors.Wrap(err, "GenerateHttpRequest")
 	}
 	return req, err
 }
@@ -51,12 +46,16 @@ func (api *CoinGeckoExecutor) ExtractRate(resp []byte, currency *models.Currency
 	var responseData CoinGeckoResponse
 	err := json.NewDecoder(bytes.NewReader(resp)).Decode(&responseData)
 	if err != nil {
-		return rate, errors.Wrap(customErrors.ErrFailedToGetRate, "extractRate")
+		return rate, errors.Wrap(ErrFailedToGetRate, "ExtractRate")
 	}
 
 	rate.Value = responseData[currenciesCoinGecko[currency.CurrencyFrom]][currenciesCoinGecko[currency.CurrencyTo]]
 	if rate.Value == 0 {
-		return rate, errors.Wrap(customErrors.ErrFailedToGetRate, "extractRate")
+		return rate, errors.Wrap(ErrFailedToGetRate, "ExtractRate")
 	}
 	return rate, nil
+}
+
+func (api *CoinGeckoExecutor) generateEndpoint(currency *models.Currency) string {
+	return fmt.Sprintf(api.url, currency.CurrencyFrom, currency.CurrencyTo)
 }
