+++ internal/currencyrate/error.go
@@ -1,4 +1,4 @@
-package customErrors
+package currencyrate
 
 import "errors"
