+++ internal/email/error.go
@@ -1,4 +1,4 @@
-package customErrors
+package email
 
 import "errors"
