+++ internal/email/templates/simple_text_templates.go
@@ -1,8 +1,8 @@
 package templates
 
 import (
-	"btc-test-task/internal/helpers/config"
-	"btc-test-task/internal/helpers/models"
+	"btc-test-task/internal/configuration/config"
+	"btc-test-task/internal/models"
 	"fmt"
 )
