+++ internal/models/email_model.go
