+++ internal/models/rate_model.go
