+++ internal/server/handlers/handlers_factory_impl.go
@@ -1,7 +1,7 @@
 package handlers
 
 import (
-	"btc-test-task/internal/helpers/config"
+	"btc-test-task/internal/configuration/config"
 )
 
 type HandlersFactoryImpl struct {
