+++ internal/server/handlers/rate.go
@@ -1,19 +1,19 @@
 package handlers
 
 import (
-	errors2 "btc-test-task/internal/helpers/customErrors"
-	"btc-test-task/internal/helpers/logger"
-	"btc-test-task/internal/helpers/models"
+	"btc-test-task/internal/configuration/logger"
+	"btc-test-task/internal/currencyrate"
+	"btc-test-task/internal/models"
 	"net/http"
 
 	"github.com/pkg/errors"
 )
 
 func (factory *HandlersFactoryImpl) CreateRate() http.HandlerFunc {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+	return func(w http.ResponseWriter, r *http.Request) {
 		rate, err := factory.services.GetRateProviderService().GetCurrentRate(
 			&models.Currency{CurrencyFrom: factory.currencyFrom, CurrencyTo: factory.currencyTo})
-		if errors.Is(err, errors2.ErrFailedToGetRate) {
+		if errors.Is(err, currencyrate.ErrFailedToGetRate) {
 			logger.Log.Warn(err)
 			w.WriteHeader(http.StatusNotFound)
 			return
@@ -30,5 +30,5 @@ func (factory *HandlersFactoryImpl) CreateRate() http.HandlerFunc {
 			return
 		}
 		w.WriteHeader(http.StatusOK)
-	})
+	}
 }
