+++ internal/server/server.go
@@ -1,7 +1,7 @@
 package server
 
 import (
-	"btc-test-task/internal/helpers/config"
+	"btc-test-task/internal/configuration/config"
 	"fmt"
 
 	"net/http"
