+++ internal/server/tests/handlers/subscribe/subscribe_test.go
@@ -1,11 +1,11 @@
 package subscribeTest
 
 import (
-	emailsStorageTest "btc-test-task/internal/emailsRepository/tests"
-	"btc-test-task/internal/helpers/config"
-	"btc-test-task/internal/helpers/logger"
-	"btc-test-task/internal/helpers/models"
-	"btc-test-task/internal/helpers/types"
+	"btc-test-task/internal/configuration/config"
+	"btc-test-task/internal/configuration/logger"
+	"btc-test-task/internal/lifecycle"
+	"btc-test-task/internal/models"
+	emailsStorageTest "btc-test-task/internal/repository/tests"
 	"btc-test-task/internal/server/handlers"
 	"bytes"
 	"fmt"
@@ -47,7 +47,7 @@ func createSubscribeRequest(body io.Reader) *http.Request {
 }
 
 func createSubscribeHandler() (http.HandlerFunc, error) {
-	servicesStubs := new(types.Services)
+	servicesStubs := new(lifecycle.Services)
 	servicesStubs.EmailsRepository = &emailsStorageTest.EmailsStorageStub{}
 	factory := handlers.NewHandlersFactoryImpl(&conf, servicesStubs)
 	subscribeHandler := factory.CreateSubscribe()
